register component moved to UI blocks
